Passes 407/407 test on Julia1.0 and 1.0.1.
Still there is a runtime error that needs to be tracked down, possibly in mcmc_algorithm.jl. Will produce a minimal working example to reproduce, although when not in test mode, the error seems to appear only during the first invocation of the mcmc_callback_vector method, only to disappear at the next attempt.